replace RealisticVisionController with RealisticVisionFly
